requirements.txt
pandas
sqlalchemy
pydantic
python-dotenv
psycopg2
